More refactorings for easier overview.
